# [precision.py] Wall-clock time speedup by performing top-k slicing at first and replacing `argsort` with `topk`