--- a/torchmetrics/functional/retrieval/precision.py
+++ b/torchmetrics/functional/retrieval/precision.py
@@ -52,10 +52,8 @@ def retrieval_precision(preds: Tensor, target: Tensor, k: Optional[int] = None) 
     if not (isinstance(k, int) and k > 0):
         raise ValueError("`k` has to be a positive integer or None")
 
-    k = min(k, preds.shape[-1])
-
     if not target.sum():
         return tensor(0.0, device=preds.device)
 
-    relevant = target[preds.topk(k, dim=-1)[1]].sum().float()
+    relevant = target[preds.topk(min(k, preds.shape[-1]), dim=-1)[1]].sum().float()
     return relevant / k