## What does this PR do?

This PR proposes two speedup improvements:
1. By performing top-k slicing before indexing the target
2. Replace `argsort` with `topk`: it seems that the latter is much more efficient than former

## Pitch

A toy example to illustrate **[1]**:

```python
a = torch.randn(1024)
b = torch.randint(0, 2, size=(1024,))

%timeit -n 1000 b[torch.argsort(a, dim=-1, descending=True)][:256]
# 139 µs ± 13 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
%timeit -n 1000 b[torch.argsort(a, dim=-1, descending=True)[:256]]
# 124 µs ± 13.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

On a single CPU, it's about 10% faster.
Since retrieval metrics are computed per example over the dataset, the overall speedup could be significant in certain cases.

A toy example to illustrate **[2]**:

```python
a = torch.randn(1024)
b = torch.randint(0, 2, size=(1024,))

%timeit -n 1000 b[torch.argsort(a, dim=-1, descending=True)[:256]]
# 114 µs ± 17.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
%timeit -n 1000 b[a.topk(256, dim=-1)[1]]
# 46.7 µs ± 4.33 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

This leads to ~**2.4X** speedup in CPU mode which can be quite significant.