@zuoxingdong 2021-11-18
> @zuoxingdong the failing tests seems to be related to the changes in the PR <img alt="Skærmbillede 2021-11-17 104103" width="591" src="https://user-images.githubusercontent.com/24896311/142176181-d2d3fd29-75f3-434b-b4dd-df154fe6e14d.png">

@SkafteNicki Thanks a lot for pointing this out!
It seems that the failures are due to an out-of-bound value of `k`. 
The existing implementation calculates the metrics simply using `k` to divide `relevant`. But I think we shouldn't do this when `k` is larger than the second dim. Instead, an upper bound cutoff is supposed to take place. 
What do you think?