@zuoxingdong 2021-11-23
@SkafteNicki Thanks a lot for your feedback :)
One tiny detail, in the changelog, I guess the order between `torch.argsort` and `torch.topk` should be flipped around:

"Use `torch.argsort` instead of `torch.topk` in retrieval precision for speedup"
-> 
"Use `torch.topk` instead of `torch.argsort` in retrieval precision for speedup"