@zuoxingdong 2021-11-24
@SkafteNicki Since it goes well with `retrieval_precision`, I'd like to work on migrating similar changes to other retrieval metrics under functional folder (e.g. `average_precision`, `fall_out`, `hit_rate`, `ndcg`, `r_precision`, `recall`, `reciprocal_rank`) as they share the same logic of speeding up via `topk`

Do you think it makes sense? If so, I'll start working on that :)