@SkafteNicki 2021-11-24
Hi @zuoxingdong,
Performance improvements are always welcome :]
Just remember to do a bit of testing (using `timeit` for example) to make sure that it actually improves the code.