@lucadiliello 2021-12-06
Thank you @zuoxingdong for your work. If you need help just contact me, I will be happy to help. There have always been a performance issue with retrieval metrics due to the many sort needed. If you can speed up things it will be very nice.